+++ v2/pkg/engine/plan/key_fields_visitor_test.go
@@ -483,7 +483,7 @@ func TestCollectKeysForPath(t *testing.T) {
 		t.Run(c.name, func(t *testing.T) {
 			definition := unsafeparser.ParseGraphqlDocumentStringWithBaseSchema(c.definition)
 
-			collectNodesVisitor := &collectNodesVisitor{
+			collectNodesVisitor := &collectNodesDSVisitor{
 				definition:          &definition,
 				dataSource:          c.dataSource,
 				providesEntries:     c.providesEntries,
@@ -492,9 +492,9 @@ func TestCollectKeysForPath(t *testing.T) {
 				notExternalKeyPaths: make(map[string]struct{}),
 			}
 
-			collectNodesVisitor.collectKeysForPath(c.typeName, c.parentPath)
+			assert.NoError(t, collectNodesVisitor.collectKeysForPath(c.typeName, c.parentPath))
 			// call it again to test the deduplication
-			collectNodesVisitor.collectKeysForPath(c.typeName, c.parentPath)
+			assert.NoError(t, collectNodesVisitor.collectKeysForPath(c.typeName, c.parentPath))
 
 			assert.Equal(t, len(c.expectKeys), len(collectNodesVisitor.keys))
 			assert.Equal(t, c.expectKeys, collectNodesVisitor.keys)