+++ v2/CHANGELOG.md
@@ -1,5 +1,47 @@
 # Changelog
 
+## [2.0.0-rc.236](https://github.com/wundergraph/graphql-go-tools/compare/v2.0.0-rc.235...v2.0.0-rc.236) (2025-10-31)
+
+
+### Bug Fixes
+
+* fix conflict of rewriter and required fields ([#1338](https://github.com/wundergraph/graphql-go-tools/issues/1338)) ([03189bc](https://github.com/wundergraph/graphql-go-tools/commit/03189bc270a551c2875b3b01dbc11d166d703e4a))
+
+## [2.0.0-rc.235](https://github.com/wundergraph/graphql-go-tools/compare/v2.0.0-rc.234...v2.0.0-rc.235) (2025-10-29)
+
+
+### Bug Fixes
+
+* improve query planning time ([#1326](https://github.com/wundergraph/graphql-go-tools/issues/1326)) ([373cf2a](https://github.com/wundergraph/graphql-go-tools/commit/373cf2abf1459ded5997e5a84d1a282dc544844c))
+
+## [2.0.0-rc.234](https://github.com/wundergraph/graphql-go-tools/compare/v2.0.0-rc.233...v2.0.0-rc.234) (2025-10-24)
+
+
+### Features
+
+* september 2025 spec description support for `OperationDocument` and `Fragment` ([#1331](https://github.com/wundergraph/graphql-go-tools/issues/1331)) ([aedaf71](https://github.com/wundergraph/graphql-go-tools/commit/aedaf7183888ee0ccd76e3baaddf44dde03df701))
+
+## [2.0.0-rc.233](https://github.com/wundergraph/graphql-go-tools/compare/v2.0.0-rc.232...v2.0.0-rc.233) (2025-10-23)
+
+
+### Bug Fixes
+
+* omit introspection queries from complexity limits ([#1332](https://github.com/wundergraph/graphql-go-tools/issues/1332)) ([971a239](https://github.com/wundergraph/graphql-go-tools/commit/971a239d0ddc8b2378f81af82f0457f8eeea0db3))
+
+## [2.0.0-rc.232](https://github.com/wundergraph/graphql-go-tools/compare/v2.0.0-rc.231...v2.0.0-rc.232) (2025-10-22)
+
+
+### Bug Fixes
+
+* correct SSE datasource complete behaviour ([#1311](https://github.com/wundergraph/graphql-go-tools/issues/1311)) ([18c39e7](https://github.com/wundergraph/graphql-go-tools/commit/18c39e7d7e42086bb710d7dc757e32c4eeed94f9))
+
+## [2.0.0-rc.231](https://github.com/wundergraph/graphql-go-tools/compare/v2.0.0-rc.230...v2.0.0-rc.231) (2025-10-20)
+
+
+### Features
+
+* support the oneOf directive ([#1308](https://github.com/wundergraph/graphql-go-tools/issues/1308)) ([251cb02](https://github.com/wundergraph/graphql-go-tools/commit/251cb029a9e232f522ab3260db3d80942222ed2c))
+
 ## [2.0.0-rc.230](https://github.com/wundergraph/graphql-go-tools/compare/v2.0.0-rc.229...v2.0.0-rc.230) (2025-10-09)